signalling-services/requirements.txt
Flask==3.1.0
flask-socketio
eventlet
flask-cors
flask-socketio==5.3.2
python-dotenv==1.0.0
PyJWT
flask-socketio 
flask-cors 
flask-sqlalchemy 
psycopg2-binary 
